requirements.txt
absl-py==2.1.0
accelerate==1.9.0
aiofiles==24.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.6.0
anthropic==0.68.0
anyascii==0.3.3
anyio==4.3.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.2.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==23.2.0
audioread==3.0.1
av==15.0.0
azure-cognitiveservices-speech==1.36.0
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.12.3
blake3==1.0.5
bleach==6.2.0
blinker==1.7.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brotli==1.1.0
cachetools==6.1.0
catalogue==2.0.10
cbor2==5.6.5
certifi==2025.8.3
cffi==1.16.0
charset-normalizer==3.4.2
click==8.1.7
cloudpathlib==0.21.1
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.3
compressed-tensors==0.10.2
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
courlan==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dataclasses-json==0.6.7
datasets==4.0.0
datasketch==1.6.5
dateparser==1.1.8
debugpy==1.8.15
decorator==5.2.1
deepface==0.0.89
defusedxml==0.7.1
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.17.0
einops==0.8.1
email_validator==2.2.0
encodec==0.1.1
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
faster-whisper==1.2.0
fastjsonschema==2.21.1
feedparser==6.0.11
ffmpy==0.6.1
filelock==3.18.0
fire==0.6.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2025.3.0
future==1.0.0
g2pkk==0.1.2
gast==0.5.4
gdown==5.1.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
gradio==5.39.0
gradio_client==1.11.0
groovy==0.1.2
grpcio==1.62.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
hangul-romanize==0.1.0
hf-xet==1.1.5
htmldate==1.9.3
httpcore==1.0.4
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.6
inflect==7.5.0
interegular==0.3.3
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
iso8601==2.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jusText==3.0.2
keras==3.1.1
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.79
langchain-ollama==0.3.6
langchain-openai==0.3.35
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.9
language_data==1.3.0
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.3
mistune==3.1.3
ml-dtypes==0.3.2
mlx==0.27.1
mlx-lm==0.26.2
mlx-metal==0.27.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
mtcnn==0.1.1
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.0.7
narwhals==2.6.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
ninja==1.11.1.4
nltk==3.9.1
notebook==7.4.4
notebook_shim==0.2.4
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
ollama==0.5.1
onnxruntime==1.22.1
openai==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-python==4.9.0.80
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
optree==0.11.0
orjson==3.11.1
outlines_core==0.2.10
overrides==7.7.0
packaging==24.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
pathlib==1.0.1
pdf2image==1.17.0
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.2.0
pooch==1.8.2
preshed==3.0.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==4.25.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.21
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.13
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.2.3
pypinyin==0.55.0
pysbd==0.3.4
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.1
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
retina-face==0.0.15
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.7.1
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
ruff==0.12.7
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.34.1
serial==0.0.97
setuptools-rust==1.11.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.2
streamlit==1.50.0
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.16.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tld==0.13.1
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
trafilatura==2.0.0
trainer==0.0.36
traitlets==5.14.3
transformers==4.54.1
TTS==0.22.0
typeguard==4.4.4
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.25.1
vllm==0.10.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.1
widgetsnbextension==4.0.14
wrapt==1.16.0
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.9.4
yt-dlp==2025.7.21
zstandard==0.23.0
